admin and other tests 🧪 fix
