+++ users/tests/test_urls.py
@@ -8,18 +8,7 @@ from users.views import followUser
 class AccountTests(APITestCase):
 
     def setUp(self):
-        url = reverse('users-api:register')
-        data = {
-            'username':'test',
-            'email':'test@gmail.com', 
-            'password':'test@123'
-        }
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(User.objects.count(), 1)
-        self.assertEqual(User.objects.get().username, 'test')
-        self.test_user = User.objects.get(username='test')
-        self.test_user_pwd = 'test@123'
+        pass
 
     def test_users_follow_url(self):
         url = 'users-api:follow-user'
